+++ lib/src/linesman_rule.dart
@@ -8,6 +8,8 @@ import 'package:analyzer/dart/ast/visitor.dart';
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/error/error.dart';
 import 'package:linesman/linesman.dart';
+import 'package:linesman/src/config.dart';
+import 'package:linesman/src/rule.dart';
 import 'package:path/path.dart' as p;
 import 'package:yaml/yaml.dart';